Review tests for from_pandas
